Extracted the bottom bar buttom generation to a `DependencyContainer` extension
